[REFACTOR] store property data in localStorage
