+++ src/components/Auth/Prefetch.jsx
@@ -1,7 +1,9 @@
 import React, { useEffect } from 'react';
 import { Outlet } from 'react-router-dom';
 import { store } from '../../store';
-import { notesApiSlice, usersApiSlice, propertiesApiSlice } from '../../store/slices';
+import { notesApiSlice, usersApiSlice, propertiesApiSlice, useGetPropertiesQuery} from '../../store/slices';
+import secureLocalStorage from 'react-secure-storage';
+import { deserializeData } from '../../utils';
 
 export const Prefetch = () => {
 
@@ -17,8 +19,27 @@ export const Prefetch = () => {
         store.dispatch(propertiesApiSlice.util.prefetch('getProperties', 'propertiesList', { force: true }))
         store.dispatch(notesApiSlice.util.prefetch('getNotes', 'notesList', { force: true }))
         store.dispatch(usersApiSlice.util.prefetch('getUsers', 'usersList', { force: true }))
-
     }, [])
+
+
+    // store properties data in LocalStorage to improve performance
+    const { properties, isLoading, isError, error } = useGetPropertiesQuery(
+        'propertiesList',
+        {
+            selectFromResult: ({ data }) => ({
+                properties: data,
+            }),
+        }
+    );
+
+    const deserializedPropertiesList = properties ? deserializeData(properties) : null;
+
+    console.log('properties List from RTK = ', properties)
+    console.log('properties Deserialized List = ', deserializedPropertiesList)
+
+    secureLocalStorage.setItem('propertiesList', deserializedPropertiesList);
+
+
 
     return <Outlet />
 };
