--- src/hooks/useDropFile/index.jsx
@@ -1,1 +1,0 @@
-export { useDropFile } from './useDropFile';
