+++ src/utils/index.js
@@ -9,3 +9,4 @@ export {
 } from './converter';
 export * from './objectUtilities';
 export { randomString } from "./randomString";
+export { deserializeData } from "./objectUtilities";
